move encoding to rrdata package
