+++ internal/dns/common/rrdata/encoding.go
@@ -1,33 +1,32 @@
-package zone
+package rrdata
 
 import (
 	"fmt"
 
-	"github.com/haukened/rr-dns/internal/dns/common/rrdata"
 	"github.com/haukened/rr-dns/internal/dns/domain"
 )
 
-// encodeRRData encodes a record value based on its type, to its binary representation.
-func encodeRRData(rrType domain.RRType, data string) ([]byte, error) {
+// EncodeRRData encodes a record value based on its type, to its binary representation.
+func EncodeRRData(rrType domain.RRType, data string) ([]byte, error) {
 	switch rrType {
 	case domain.RRTypeA: // 1
-		return rrdata.EncodeAData(data)
+		return EncodeAData(data)
 	case domain.RRTypeNS: // 2
-		return rrdata.EncodeNSData(data)
+		return EncodeNSData(data)
 	case domain.RRTypeCNAME: // 5
-		return rrdata.EncodeCNAMEData(data)
+		return EncodeCNAMEData(data)
 	case domain.RRTypeSOA: // 6
-		return rrdata.EncodeSOAData(data)
+		return EncodeSOAData(data)
 	case domain.RRTypePTR: // 12
-		return rrdata.EncodePTRData(data)
+		return EncodePTRData(data)
 	case domain.RRTypeMX: // 15
-		return rrdata.EncodeMXData(data)
+		return EncodeMXData(data)
 	case domain.RRTypeTXT: // 16
-		return rrdata.EncodeTXTData(data)
+		return EncodeTXTData(data)
 	case domain.RRTypeAAAA: // 28
-		return rrdata.EncodeAAAAData(data)
+		return EncodeAAAAData(data)
 	case domain.RRTypeSRV: // 33
-		return rrdata.EncodeSRVData(data)
+		return EncodeSRVData(data)
 	case domain.RRTypeNAPTR: // 35
 		return notimp(domain.RRTypeNAPTR)
 	case domain.RRTypeOPT: // 41
@@ -47,7 +46,7 @@ func encodeRRData(rrType domain.RRType, data string) ([]byte, error) {
 	case domain.RRTypeHTTPS: // 65
 		return notimp(domain.RRTypeHTTPS)
 	case domain.RRTypeCAA: // 257
-		return rrdata.EncodeCAAData(data)
+		return EncodeCAAData(data)
 	default:
 		return []byte(data), nil
 	}
